+++ Logos/src/utopia/logos/database/access/many/url/link/DbLinks.scala
@@ -10,6 +10,7 @@ import utopia.flow.util.StringExtensions._
 import utopia.logos.database.access.many.url.domain.DbDomains
 import utopia.logos.database.access.many.url.path.DbRequestPaths
 import utopia.logos.database.storable.url.RequestPathDbModel
+import utopia.logos.model.cached.Link
 import utopia.logos.model.combined.url.DetailedLink
 import utopia.logos.model.partial.url.{LinkData, RequestPathData}
 import utopia.logos.model.stored.url.{Domain, StoredLink}
@@ -50,7 +51,7 @@ object DbLinks extends ManyLinksAccess with UnconditionalView with ViewManyByInt
 					// Removes the / from domain, if present
 					val domainPart = link.slice(domainRange).notEndingWith("/")
 					val remainingPart = link.drop(domainRange.last + 1)
-					StoredLink.paramPartRegex.firstRangeFrom(remainingPart) match {
+					Link.paramPartRegex.firstRangeFrom(remainingPart) match {
 						// Case: Parameters are specified => Extracts them from the path
 						case Some(paramsRange) =>
 							val paramsPart = remainingPart.slice(paramsRange).drop(1)