+++ Flow/Changes.md
@@ -2,6 +2,8 @@
 
 ## v2.5.1 (in development)
 ### Deprecations
+- Moved **AsyncCollectionExtensions** features to **CollectionExtensions** 
+  because some compilers couldn't handle cases where both were imported
 - Renamed **DateRange**'s `.period` to `.length`
 ### Bugfixes
 - Previously **Path**`.hasSameContentAs(Path)` would throw if the other path was non-existent
@@ -12,8 +14,8 @@
 - Added **RepeatOneForeverIterator**
 - Added **LazilyInitializedChanging** that provides lazy **Changing** wrapping
 ### New methods
-- **Iterable** (**AsyncExtensions**)
-  - Added a new `.mapParallel(...)` variant
+- **Iterable** (**CollectionExtensions**)
+  - Added `.mapParallelUsing(ActionQueue)(...)`
 - **IterableOnce** (**CollectionExtensions**)
   - Added `.splitMapInto(...)`, `.splitFlatMapInto(...)` and `.splitInto(...)`
 - **ModelLike**